# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Opsy Backend Backend (in GoLang)",
        "title": "Opsy Backend",
        "contact": {
            "name": "Priyanshu Rai"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.0.1"
    },
    "host": "nssdmvtpj2.us-east-1.awsapprunner.com",
    "basePath": "/",
    "paths": {
...
        "/user/create-log-entry": {
            "post": {
                "description": "CreateLogEntry",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logEntry"
                ],
                "summary": "CreateLogEntry",
                "parameters": [
                    {
                        "description": "CreateLogEntry for user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logEntry.LogEntryReqDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/logEntry.GetLogEntryResDto"
                        }
                    }
                }
            }
        },
...
        "/user/edit-appointment/{id}": {
            "put": {
                "description": "Update appointment",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logEntry"
                ],
                "summary": "Update appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "alert",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "notes",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "whatItIsFor",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "when",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/logEntry.GetAppointmentResDto"
                        }
                    }
                }
            }
        },
...
        "/user/monthly-insights": {
            "get": {
                "description": "Retrieves mental health and physical health data for a specified month",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "logEntry"
                ],
                "summary": "Fetch monthly insights data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Month and Year (MM-YYYY)",
                        "name": "monthYear",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/logEntry.InsightsResDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "logEntry.AppointmentRes": {
            "type": "object",
            "properties": {
                "alert": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "whatItIsFor": {
                    "type": "string"
                },
                "when": {
                    "type": "string"
                }
            }
        },
...
        "logEntry.GetAppointmentResDto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logEntry.AppointmentRes"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "logEntry.GetLogEntryResDto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logEntry.LogEntryRes"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "logEntry.InsightsData": {
            "type": "object",
            "properties": {
                "mentalHealth": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/logEntry.MentalHealthRes"
                    }
                },
                "physicalHealth": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/logEntry.PhysicalHealthRes"
                    }
                }
            }
        },
        "logEntry.InsightsResDto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/logEntry.InsightsData"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "logEntry.LogEntryReqDto": {
            "type": "object",
            "properties": {
                "alert": {
                    "type": "string"
                },
                "dateTime": {
                    "type": "string"
                },
                "feel": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "painLevel": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "ways": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "whatItIsFor": {
                    "type": "string"
                }
            }
        },
        "logEntry.LogEntryRes": {
            "type": "object",
            "properties": {
                "alert": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "feel": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "painLevel": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "ways": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "whatItIsFor": {
                    "type": "string"
                },
                "when": {
                    "type": "string"
                }
            }
        },
...
        "logEntry.MentalHealthRes": {
            "type": "object",
            "properties": {
                "avgFeel": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "day": {
                    "type": "string"
                }
            }
        },
        "logEntry.PhysicalHealthRes": {
            "type": "object",
            "properties": {
                "avgPainLevel": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "day": {
                    "type": "string"
                }
            }
        },
...
    }
}